Opdelt EjendomsLogik i to pakker
