+++ src/main/java/spillogik/BevaegelsesLogik.java
@@ -71,4 +71,5 @@ public class BevaegelsesLogik {
 
 
 
+
 }
