+++ src/test/java/FordelingAfHuseTest.java
@@ -1,7 +1,7 @@
 import model.felter.ejeligefelter.Ejendomsgruppe;
 import model.felter.ejeligefelter.Ejendom;
 import org.junit.jupiter.api.Test;
-import spillogik.EjendomsLogik;
+import spillogik.HusHotelLogik;
 
 import java.awt.*;
 
@@ -41,27 +41,27 @@ class FordelingAfHuseTest {
         ejendom3.setGruppe(ejendomsGruppe);
 
         // Test 1 -  Ingen huse på nogen ejendomme
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom2) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom3) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom2) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom3) );
 
         // Test 2 -  Ejendom1: 1 hus
         ejendom1.setAntalHuse(1);
-        assertFalse( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom2) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom3) );
+        assertFalse( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom2) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom3) );
 
         // Test 3 - Ejendom1: 1 hus, Ejendom2: 1 hus
         ejendom2.setAntalHuse(1);
-        assertFalse( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
-        assertFalse( EjendomsLogik.huseErFordeltIGruppe(ejendom2) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom3) );
+        assertFalse( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
+        assertFalse( HusHotelLogik.huseErFordeltIGruppe(ejendom2) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom3) );
 
         // Test 4 - Alle ejendomme: 1 hus
         ejendom3.setAntalHuse(1);
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom2) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom3) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom2) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom3) );
 
         // HOTELLER -------
 
@@ -72,7 +72,7 @@ class FordelingAfHuseTest {
         ejendom3.setAntalHuse(0);
         ejendom2.setHarHotel(true);
 
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
 
 
         // Test 6 - Man skal ikke kunne bygge på ejendom1
@@ -83,7 +83,7 @@ class FordelingAfHuseTest {
         ejendom3.setAntalHuse(3);
         ejendom2.setHarHotel(false);
 
-        assertFalse( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
+        assertFalse( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
 
 
         // Test 7 - Man skal ikke kunne bygge på alle ejendomme
@@ -96,9 +96,9 @@ class FordelingAfHuseTest {
         ejendom3.setAntalHuse(4);
         ejendom2.setHarHotel(false);
 
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
-        assertTrue( EjendomsLogik.huseErFordeltIGruppe(ejendom1) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
+        assertTrue( HusHotelLogik.huseErFordeltIGruppe(ejendom1) );
 
 
     }
